+++ AutomatedLab/AutomatedLabVirtualMachines.psm1
@@ -2419,8 +2419,10 @@ function Copy-LabALCommon
         [string[]]
         $ComputerName
     )
+
+    $machines = Get-LabVm -ComputerName $ComputerName | Where-Object {-not $_.InternalNotes.ALCommonCopied}
     
-    $childPath = foreach ($vm in $ComputerName)
+    $childPath = foreach ($vm in $machines)
     {
         Invoke-LabCommand -ScriptBlock {
             if ($PSEdition -eq 'Core')
@@ -2431,27 +2433,34 @@ function Copy-LabALCommon
                 'full'
             }
         } -ComputerName $vm -NoDisplay -IgnoreAzureLabSources -PassThru |
-        Add-Member -MemberType NoteProperty -Name ComputerName -Value $vm -Force -PassThru
+        Add-Member -MemberType NoteProperty -Name ComputerName -Value $vm.Name -Force -PassThru
     }
 
     $coreChild = @($childPath) -eq 'core'
     $fullChild = @($childPath) -eq 'full'
     $libLocation = Split-Path -Parent -Path (Split-Path -Path ([AutomatedLab.Common.Win32Exception]).Assembly.Location -Parent)
 
-    if ($coreChild -and @(Invoke-LabCommand -ScriptBlock{
-                Get-Item -Path '/ALLibraries/core/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-    } -ComputerName $coreChild.ComputerName -IgnoreAzureLabSources -NoDisplay -PassThru).Count -ne $coreChild.Count)
+    if ($coreChild)
     {
         $coreLibraryFolder = Join-Path -Path $libLocation -ChildPath $coreChild[0]
         Copy-LabFileItem -Path $coreLibraryFolder -ComputerName $coreChild.ComputerName -DestinationFolderPath '/ALLibraries' -UseAzureLabSourcesOnAzureVm $false
     }
 
-    if ($fullChild -and @(Invoke-LabCommand -ScriptBlock {
-                Get-Item -Path '/ALLibraries/full/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-    } -ComputerName $fullChild.ComputerName -IgnoreAzureLabSources -NoDisplay -PassThru).Count -ne $fullChild.Count)
+    if ($fullChild)
     {
         $fullLibraryFolder = Join-Path -Path $libLocation -ChildPath $fullChild[0]
         Copy-LabFileItem -Path $fullLibraryFolder -ComputerName $fullChild.ComputerName -DestinationFolderPath '/ALLibraries' -UseAzureLabSourcesOnAzureVm $false
+    }
+
+    foreach ($machine in $machines)
+    {
+        $shouldExport = $true
+        $machine.InternalNotes.Add('ALCommonCopied', $true)
+    }
+
+    if ($shouldExport)
+    {
+        Export-Lab
     }
 }
 #endregion Copy-LabALCommon